+++ resources/views/welcome.blade.php
@@ -73,6 +73,16 @@
             </div>
         @endallowed
 
+        <div class="card mb-4">
+            <div class="card-header">
+                Changelog
+                <span class="pull-right">Version: <strong>{{ $app_version }}</strong></span>
+            </div>
+            <div class="card-body">
+                <p>Read about the latest changes of this application <a href="{{ route('changelog') }}">here</a>.</p>
+            </div>
+        </div>
+
     </div>
 
 @endsection