+++ Changelog.md
@@ -4,3 +4,4 @@
 
 * Added view to read the changelog
 * Persons will now be marked as "frequent visitor" based on their number of visits at the bank. Threshold can be defined in bank settings.
+* People report has been enhanced, showing important numbers at the top of the page