+++ Changelog.md
@@ -2,6 +2,6 @@
 
 ## 1.8.17
 
-* Added view to read the changelog
+* Added view to read the changelog. Changelog card has been added to the dashboard.
 * Persons will now be marked as "frequent visitor" based on their number of visits at the bank. Threshold can be defined in bank settings.
-* People report has been enhanced, showing important numbers at the top of the page
+* People report has been enhanced, showing important numbers at the top of the page.